packages/python-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tanzil-core"
version = "0.1.0"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.8.0",
    "python-statemachine>=2.0.0"
]

[project.entry-points."tanzil.components"]
downloader = "tanzil.components.downloader.manager:DownloadManager"

[tool.setuptools.packages.find]
where = ["."]
include = ["tanzil*"]
